[CHECKPOINT] on misc fixes, improve layout, colors and stability
